Desupport perl, ruby bindings-- move them to /contrib
Summary:
Moved the Perl and Ruby bindings to an attic-type area
called "contrib" to mark them as not-officially supported parts
of the SDK distribution. The idea is that if someone (either a
DNAnexan or not) wants to revive it at some point, they would
have a useful starting point to consult, but otherwise this code
isn't presumed to have any guarantees and people shouldn't be
thinking about "will I break this" if they want to make other
changes.

* Stop building the corresponding Debian packages too

Test Plan:
* Ran make targets by hand
* Verified that no applets rely on the deb packages in their execDepends

Reviewers: gduggal

Reviewed By: gduggal

Differential Revision: https://dnanexus.phacility.com/D1224