Don't bother freeing "data" to save memory
For file uploads, `jsonify_data` is false, so `serialized_data` and
`data` are aliased, and there's no savings from freeing `data`. For
other API calls, it is possible that this code is invoked like

```
dxpy.api.record_set_details(record_id, read_file_to_get_details_as_str_containing_json(...), ...)
```

where the string could be freed as soon as we had jsonified it. But it's
probably pretty rare that the size of the payload here would be large
enough to matter.