DEVEX-1 - Fix 'dx get' untar issue
Summary: The Python tarfile.extractall procedure doesn't behave quite the same as 'tar -C' in the sense that it doesn't strip the leading '/' from a path that begins with a '/'.  This implemention defines a procedure that behaves similarly using 'tarfile' while still allowing execution of 'dx get' from a Windows machine.

Test Plan: Running Python integration tests in the toolkit

Reviewers: psung

Subscribers: nicolasbockg