DEVEX-74 Fix dx-docker working directory default argument
Summary: This commmit fixes an issue where the default workdir is set to "/" in the argument parsing.  However, due to a fix introduced by 28bfd33d7a85b265fab9e904b803df167bc4c6e2, setting the default workdir in the argument parsing is not the correct behavior.  The fix is simply to remove the default from the argument parsing.

Test Plan: Ran dx-docker test suite.  Added a test to ensure default working directory is read properly and output as a part of metadata.

Reviewers: nicolasbockg

Reviewed By: nicolasbockg

Subscribers: mlin, psung

Differential Revision: https://dnanexus.phacility.com/D1283