Mark routes as retryable when a nonce is supplied
Routes that supply a nonce should be retryable (this being the whole
point of adding the nonce). We don't do this by marking such routes as
retryable in the wrapper_table.json, since not all the language bindings
know how to generate the nonces, and this could result in bindings that
didn't supply the nonce but went ahead and considered the route to be
retryable anyway. Therefore, the *bindings generators* for each language
are responsible for marking each route as retryable if and only if they
inserted nonce generation code.

In particular, note that in the Java bindings, each API route has
multiple overloads and we added nonce support in only some of them (not
bothering to add nonce support in the older and deprecated overloads).
In this case, *only* the overloads that specifically support nonces also
mark the request as retryable.