DEVEX-176-Fix-Download-All-Inputs
Summary:
A recent change to download_all_inputs caused a regression where the max number
of threads was set to None. The solution was to move the default value from the wrapper to the
actual function.

CHANGELOG:
Fixed an issue where download all inputs in parallel would hang indefintely.

Test Plan: Current tests

Reviewers: kjensen

Reviewed By: kjensen

Subscribers: gduggal

Differential Revision: https://dnanexus.phacility.com/D1331