DEVEX-204: Add alternative export mechanism to address potential bug in Docker engine itself
Summary: On a specific Docker image, we noticed that the filesystem is mangled after running 'docker save' or 'docker export'. While digging into this a little deeper and potentially pursuing the issue with folks at Docker, here is an alternative mechanism for export that is meant only to be used in extreme situations like this.

Test Plan:
Tested alternative export manually in a variety of situations based on ubuntu:14.04 or ubuntu:latest images.
Ran integration tests to make sure existing functionality is not affected negatively.
Not adding tests specific to this in the hopes that the core Docker issue is resolved and this will be unncessary.

Reviewers: azalcman

Reviewed By: azalcman

Subscribers: kjensen

Differential Revision: https://dnanexus.phacility.com/D1369