DEVEX-238: Add client support for project-based resources for multi-region apps
Summary:
Previously when building an app, a developer could specify additional
data objects that should be in the app's resources container. This feature is
now added for multi-region apps. The developer can set resources in the
"regionalOptions" field by region. The deprecated top-level "resources"
can still be used with a single-region app when "regionalOptions" is not set.

Test Plan: added integration tests

Reviewers: pnguyen

Reviewed By: pnguyen

Subscribers: psung

Differential Revision: https://dnanexus.phacility.com/D1452