Fixing bug with asset in dxapp.json
If an asset is specified in dxapp.json using a project and path, then we should require
1. That the asset exist in the exact path specified (i.e. not look in subfolders for the given path)
2. Insist that there be only a single asset that matches the description.

Currently, it could be that a user specifies an asset like
project=project-982uq3954732894324
folder=/foo
name=my_asset

Inside of project-982uq3954732894324 there could be an asset called my_asset in the /foo folder and
also an asset in the /foo/.archive folder.  The system may return the asset from the .archive folder
and the user would be none the wiser!  This seems bad.  If a user specifies the path to an asset,
we should require it be that exact path.