PTFM-21231: Inline cluster bootstrapScript (#324)
'clusterSpec' in dxapp.json can contain an optional 'bootstrapScript' field provided as a relative 
file path. During the app(let) build process, we read the file at the provided path and inline that file's 
code directly in the app(let) doc. We also write the 'bootstrapScript' code back into a file as part of 'dx get'.