PTFM-21023 dx watch job log relay test
This commit adds a test for dx watch to test bad
auth scenarios and make sure errors are communicated
correctly from the apiserver route.
A recent change to the apiserver changes the error code
for closing the websocket from 1008 to 4008.
This test makes sure that despite that socket closing
code being different, the error reasons still look
the same to the client.