Wait file list (#352)
Adding ability to specify a local file which contains a list of file objects
or jobs to wait on.   If a user has a lot of objects on which to wait, they may hit
bash limits on the number of characters able to provide as input on the command
line. This allows them to get around that.