Allow newer versions of psutil
I think this is the last one! The version of psutil is from late 2015, which again is old enough that it hits version conflicts when attempting to install `dxpy` alongside other python packages.  Any possibility this could also be allowed to float versions?

(I've put these up in separate PRs in case you only want to accept some and not others, etc)