Removing use of ubuntu 12.04 in test_dxclient.py (#441)
Using ubuntu 14.04 instead of 12.04, to run the "sleep" applet. 
 
Upgrading the dx_ssh_test to use ubuntu 14.04 instead of 12.04. It is used to run a `sleep` applet, and log into it with ssh. The shell behaves differently, for example, color codes the characters, so the previous pexpect tests were failing.