DEVEX-1591: add dx-mount-all-inputs script (similar to dx-download-all-inputs) (#603)
* DEVEX-1591: dx-mount-all-inputs script, which is similar to
dx-download-all-inputs except that instead of downloading
the input files, it mounts them using dxfuse. This helps
the performance of apps like csv_loader which deal with
very large input files.

* DEVEX-1591: per review, make sure mounted locations exactly match
downloaded locations, e.g. /home/dnanexus/in/csv/* instead of
/home/dnanexus/mount/in/csv/*.

* DEVEX-1519: proper check for dxfuse command on the platform where the mount
script is running, as is done elsewhere in the toolkit when looking for
external programs like wget.

* DEVEX-1591: minor cleanup from review

* DEVEX-1591: added app test for dx-mount-all-inputs

* DEVEX-1591 Use always the latest dxfuse version

* Create resource dir if neeeded

* Change permissions on dxfuse

* Revert "DEVEX-1591 Use always the latest dxfuse version"

This reverts commit 765c96f46d862ef197af61375950a0c3a1a30391.

* Revert "Create resource dir if neeeded"

This reverts commit fcf9c4c33d664c339d1065f3cdc892fc30e87a2b.

Reverting to check a failing test separately

* DEVEX-1591: per review feedback, print dxfuse version being used
and configure test app version props

Co-authored-by: Aleksandra Zalcman <biokere@gmail.com>