Remove very old/optional dependencies
1. Remove windows-only colorama dependency (that seems like it doesn't
matter regardless?). This version of colorama is from 2014. Makes dxpy
unable to work with other packages.
2. Remove pyreadline and gnureadline dependency (both repos say
generally not needed) - codebase is designed to not have readline
installed and still work.

While readline may still be necessary on OSX, dxpy prints out that it's
not available and all usages of readline are within try/except so it
should not actually break anything.